backend/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# AI Integration
google-generativeai==0.3.2

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
filetype==1.2.0

# Environment
python-dotenv==1.0.0

# Async support
aiofiles==23.2.1
